Clean up crypto store SQL queries

+++ mautrix/crypto/store/asyncpg/store.py
@@ -33,9 +33,12 @@ from ..abstract import CryptoStore, StateStore
 from .upgrade import upgrade_table
 
 try:
+    from sqlite3 import sqlite_version_info as sqlite_version
+
     from aiosqlite import Cursor
 except ImportError:
     Cursor = None
+    sqlite_version = (0, 0, 0)
 
 
 class PgCryptoStateStore(PgStateStore, StateStore):
@@ -75,43 +78,38 @@ class PgCryptoStore(CryptoStore, SyncStore):
                 await conn.execute(f"DELETE FROM {table} WHERE account_id=$1", self.account_id)
 
     async def get_device_id(self) -> DeviceID | None:
-        device_id = await self.db.fetchval(
-            "SELECT device_id FROM crypto_account WHERE account_id=$1", self.account_id
-        )
+        q = "SELECT device_id FROM crypto_account WHERE account_id=$1"
+        device_id = await self.db.fetchval(q, self.account_id)
         self._device_id = device_id or self._device_id
         return self._device_id
 
     async def put_device_id(self, device_id: DeviceID) -> None:
-        await self.db.fetchval(
-            "UPDATE crypto_account SET device_id=$1 WHERE account_id=$2",
-            device_id,
-            self.account_id,
-        )
+        q = "UPDATE crypto_account SET device_id=$1 WHERE account_id=$2"
+        await self.db.fetchval(q, device_id, self.account_id)
         self._device_id = device_id
 
     async def put_next_batch(self, next_batch: SyncToken) -> None:
         self._sync_token = next_batch
-        await self.db.execute(
-            "UPDATE crypto_account SET sync_token=$1 WHERE account_id=$2",
-            self._sync_token,
-            self.account_id,
-        )
+        q = "UPDATE crypto_account SET sync_token=$1 WHERE account_id=$2"
+        await self.db.execute(q, self._sync_token, self.account_id)
 
     async def get_next_batch(self) -> SyncToken:
         if self._sync_token is None:
-            self._sync_token = await self.db.fetchval(
-                "SELECT sync_token FROM crypto_account WHERE account_id=$1", self.account_id
-            )
+            q = "SELECT sync_token FROM crypto_account WHERE account_id=$1"
+            self._sync_token = await self.db.fetchval(q, self.account_id)
         return self._sync_token
 
     async def put_account(self, account: OlmAccount) -> None:
         self._account = account
         pickle = account.pickle(self.pickle_key)
+        q = """
+        INSERT INTO crypto_account (account_id, device_id, shared, sync_token, account)
+        VALUES ($1, $2, $3, $4, $5)
+        ON CONFLICT (account_id) DO UPDATE
+            SET shared=excluded.shared, sync_token=excluded.sync_token, account=excluded.account
+        """
         await self.db.execute(
-            "INSERT INTO crypto_account (account_id, device_id, shared, "
-            "sync_token, account) VALUES($1, $2, $3, $4, $5) "
-            "ON CONFLICT (account_id) DO UPDATE SET shared=$3, sync_token=$4,"
-            "                                       account=$5",
+            q,
             self.account_id,
             self._device_id,
             account.shared,
@@ -121,10 +119,8 @@ class PgCryptoStore(CryptoStore, SyncStore):
 
     async def get_account(self) -> OlmAccount:
         if self._account is None:
-            row = await self.db.fetchrow(
-                "SELECT shared, account, device_id FROM crypto_account WHERE account_id=$1",
-                self.account_id,
-            )
+            q = "SELECT shared, account, device_id FROM crypto_account WHERE account_id=$1"
+            row = await self.db.fetchrow(q, self.account_id)
             if row is not None:
                 self._account = OlmAccount.from_pickle(
                     row["account"], passphrase=self.pickle_key, shared=row["shared"]
@@ -134,19 +130,16 @@ class PgCryptoStore(CryptoStore, SyncStore):
     async def has_session(self, key: IdentityKey) -> bool:
         if len(self._olm_cache[key]) > 0:
             return True
-        val = await self.db.fetchval(
-            "SELECT session_id FROM crypto_olm_session WHERE sender_key=$1 AND account_id=$2",
-            key,
-            self.account_id,
-        )
+        q = "SELECT session_id FROM crypto_olm_session WHERE sender_key=$1 AND account_id=$2"
+        val = await self.db.fetchval(q, key, self.account_id)
         return val is not None
 
     async def get_sessions(self, key: IdentityKey) -> list[Session]:
-        q = (
-            "SELECT session_id, session, created_at, last_encrypted, last_decrypted "
-            "FROM crypto_olm_session WHERE sender_key=$1 AND account_id=$2 "
-            "ORDER BY last_decrypted DESC"
-        )
+        q = """
+        SELECT session_id, session, created_at, last_encrypted, last_decrypted
+        FROM crypto_olm_session WHERE sender_key=$1 AND account_id=$2
+        ORDER BY last_decrypted DESC
+        """
         rows = await self.db.fetch(q, key, self.account_id)
         sessions = []
         for row in rows:
@@ -165,11 +158,11 @@ class PgCryptoStore(CryptoStore, SyncStore):
         return sessions
 
     async def get_latest_session(self, key: IdentityKey) -> Session | None:
-        q = (
-            "SELECT session_id, session, created_at, last_encrypted, last_decrypted "
-            "FROM crypto_olm_session WHERE sender_key=$1 AND account_id=$2 "
-            "ORDER BY last_decrypted DESC LIMIT 1"
-        )
+        q = """
+        SELECT session_id, session, created_at, last_encrypted, last_decrypted
+        FROM crypto_olm_session WHERE sender_key=$1 AND account_id=$2
+        ORDER BY last_decrypted DESC LIMIT 1
+        """
         row = await self.db.fetchrow(q, key, self.account_id)
         if row is None:
             return None
@@ -192,9 +185,9 @@ class PgCryptoStore(CryptoStore, SyncStore):
         self._olm_cache[key][SessionID(session.id)] = session
         pickle = session.pickle(self.pickle_key)
         q = """
-            INSERT INTO crypto_olm_session (session_id, sender_key, session, created_at,
-                                            last_encrypted, last_decrypted, account_id)
-            VALUES ($1, $2, $3, $4, $5, $6, $7)
+        INSERT INTO crypto_olm_session (
+            session_id, sender_key, session, created_at, last_encrypted, last_decrypted, account_id
+        ) VALUES ($1, $2, $3, $4, $5, $6, $7)
         """
         await self.db.execute(
             q,
@@ -216,10 +209,10 @@ class PgCryptoStore(CryptoStore, SyncStore):
                 f"isn't equal to the one being saved to the database ({e})"
             )
         pickle = session.pickle(self.pickle_key)
-        q = (
-            "UPDATE crypto_olm_session SET session=$1, last_encrypted=$2, last_decrypted=$3 "
-            "WHERE session_id=$4 AND account_id=$5"
-        )
+        q = """
+        UPDATE crypto_olm_session SET session=$1, last_encrypted=$2, last_decrypted=$3
+        WHERE session_id=$4 AND account_id=$5
+        """
         await self.db.execute(
             q, pickle, session.last_encrypted, session.last_decrypted, session.id, self.account_id
         )
@@ -233,10 +226,13 @@ class PgCryptoStore(CryptoStore, SyncStore):
     ) -> None:
         pickle = session.pickle(self.pickle_key)
         forwarding_chains = ",".join(session.forwarding_chain)
+        q = """
+        INSERT INTO crypto_megolm_inbound_session (
+            session_id, sender_key, signing_key, room_id, session, forwarding_chains, account_id
+        ) VALUES ($1, $2, $3, $4, $5, $6, $7)
+        """
         await self.db.execute(
-            "INSERT INTO crypto_megolm_inbound_session (session_id, sender_key, "
-            "signing_key, room_id, session, forwarding_chains, account_id) "
-            "VALUES ($1, $2, $3, $4, $5, $6, $7)",
+            q,
             session_id,
             sender_key,
             session.signing_key,
@@ -268,13 +264,11 @@ class PgCryptoStore(CryptoStore, SyncStore):
         )
 
     async def has_group_session(self, room_id: RoomID, session_id: SessionID) -> bool:
-        count = await self.db.fetchval(
-            "SELECT COUNT(session) FROM crypto_megolm_inbound_session "
-            "WHERE room_id=$1 AND session_id=$2 AND account_id=$3",
-            room_id,
-            session_id,
-            self.account_id,
-        )
+        q = """
+        SELECT COUNT(session) FROM crypto_megolm_inbound_session
+        WHERE room_id=$1 AND session_id=$2 AND account_id=$3
+        """
+        count = await self.db.fetchval(q, room_id, session_id, self.account_id)
         return count > 0
 
     async def add_outbound_group_session(self, session: OutboundGroupSession) -> None:
@@ -282,12 +276,18 @@ class PgCryptoStore(CryptoStore, SyncStore):
         max_age = session.max_age
         if self.db.scheme == Scheme.SQLITE:
             max_age = max_age.total_seconds()
+        q = """
+        INSERT INTO crypto_megolm_outbound_session (
+            room_id, session_id, session, shared, max_messages, message_count,
+            max_age, created_at, last_used, account_id
+        ) VALUES ($1, $2, $3, $4, $5, $6, $7, $8, $9, $10)
+        ON CONFLICT (account_id, room_id) DO UPDATE
+        SET session_id=excluded.session_id, session=excluded.session, shared=excluded.shared,
+            max_messages=excluded.max_messages, message_count=excluded.message_count,
+            max_age=excluded.max_age, created_at=excluded.created_at, last_used=excluded.last_used
+        """
         await self.db.execute(
-            "INSERT INTO crypto_megolm_outbound_session (room_id, session_id, session, shared, "
-            "max_messages, message_count, max_age, created_at, last_used, account_id) "
-            "VALUES ($1, $2, $3, $4, $5, $6, $7, $8, $9, $10)"
-            "ON CONFLICT (account_id, room_id) DO UPDATE SET session_id=$2, session=$3, shared=$4,"
-            " max_messages=$5, message_count=$6, max_age=$7, created_at=$8, last_used=$9",
+            q,
             session.room_id,
             session.id,
             pickle,
@@ -302,9 +302,12 @@ class PgCryptoStore(CryptoStore, SyncStore):
 
     async def update_outbound_group_session(self, session: OutboundGroupSession) -> None:
         pickle = session.pickle(self.pickle_key)
+        q = """
+        UPDATE crypto_megolm_outbound_session SET session=$1, message_count=$2, last_used=$3
+        WHERE room_id=$4 AND session_id=$5 AND account_id=$6
+        """
         await self.db.execute(
-            "UPDATE crypto_megolm_outbound_session SET session=$1, message_count=$2, last_used=$3 "
-            "WHERE room_id=$4 AND session_id=$5 AND account_id=$6",
+            q,
             pickle,
             session.message_count,
             session.use_time,
@@ -314,13 +317,12 @@ class PgCryptoStore(CryptoStore, SyncStore):
         )
 
     async def get_outbound_group_session(self, room_id: RoomID) -> OutboundGroupSession | None:
-        row = await self.db.fetchrow(
-            "SELECT room_id, session_id, session, shared, max_messages, message_count, max_age, "
-            "       created_at, last_used "
-            "FROM crypto_megolm_outbound_session WHERE room_id=$1 AND account_id=$2",
-            room_id,
-            self.account_id,
-        )
+        q = """
+        SELECT room_id, session_id, session, shared, max_messages, message_count, max_age,
+               created_at, last_used
+        FROM crypto_megolm_outbound_session WHERE room_id=$1 AND account_id=$2
+        """
+        row = await self.db.fetchrow(q, room_id, self.account_id)
         if row is None:
             return None
         max_age = row["max_age"]
@@ -339,39 +341,29 @@ class PgCryptoStore(CryptoStore, SyncStore):
         )
 
     async def remove_outbound_group_session(self, room_id: RoomID) -> None:
-        await self.db.execute(
-            "DELETE FROM crypto_megolm_outbound_session WHERE room_id=$1 AND account_id=$2",
-            room_id,
-            self.account_id,
-        )
+        q = "DELETE FROM crypto_megolm_outbound_session WHERE room_id=$1 AND account_id=$2"
+        await self.db.execute(q, room_id, self.account_id)
 
     async def remove_outbound_group_sessions(self, rooms: list[RoomID]) -> None:
         if self.db.scheme in (Scheme.POSTGRES, Scheme.COCKROACH):
-            await self.db.execute(
-                "DELETE FROM crypto_megolm_outbound_session "
-                "WHERE account_id=$1 AND room_id=ANY($2)",
-                self.account_id,
-                rooms,
-            )
+            q = """
+            DELETE FROM crypto_megolm_outbound_session WHERE account_id=$1 AND room_id=ANY($2)
+            """
+            await self.db.execute(q, self.account_id, rooms)
         else:
             params = ",".join(["?"] * len(rooms))
-            await self.db.execute(
-                "DELETE FROM crypto_megolm_outbound_session "
-                f"WHERE account_id=? AND room_id IN ({params})",
-                self.account_id,
-                *rooms,
-            )
+            q = f"""
+            DELETE FROM crypto_megolm_outbound_session WHERE account_id=? AND room_id IN ({params})
+            """
+            await self.db.execute(q, self.account_id, *rooms)
 
-    _validate_message_index_query = (
-        "WITH existing AS ("
-        " INSERT INTO crypto_message_index(sender_key, session_id, index, event_id, timestamp)"
-        " VALUES ($1, $2, $3, $4, $5)"
-        # have to update something so that RETURNING * always returns the row
-        " ON CONFLICT (sender_key, session_id, index) DO UPDATE SET sender_key=$1"
-        " RETURNING *"
-        ")"
-        "SELECT * FROM existing"
-    )
+    _validate_message_index_query = """
+    INSERT INTO crypto_message_index (sender_key, session_id, index, event_id, timestamp)
+    VALUES ($1, $2, $3, $4, $5)
+    -- have to update something so that RETURNING * always returns the row
+    ON CONFLICT (sender_key, session_id, index) DO UPDATE SET sender_key=excluded.sender_key
+    RETURNING *
+    """
 
     async def validate_message_index(
         self,
@@ -381,7 +373,11 @@ class PgCryptoStore(CryptoStore, SyncStore):
         index: int,
         timestamp: int,
     ) -> bool:
-        if self.db.scheme in (Scheme.POSTGRES, Scheme.COCKROACH):
+        if self.db.scheme in (Scheme.POSTGRES, Scheme.COCKROACH) or (
+            # RETURNING was added in SQLite 3.35.0 https://www.sqlite.org/lang_returning.html
+            self.db.scheme == Scheme.SQLITE
+            and sqlite_version >= (3, 35)
+        ):
             row = await self.db.fetchrow(
                 self._validate_message_index_query,
                 sender_key,
@@ -414,16 +410,15 @@ class PgCryptoStore(CryptoStore, SyncStore):
             return True
 
     async def get_devices(self, user_id: UserID) -> dict[DeviceID, DeviceIdentity] | None:
-        tracked_user_id = await self.db.fetchval(
-            "SELECT user_id FROM crypto_tracked_user WHERE user_id=$1", user_id
-        )
+        q = "SELECT user_id FROM crypto_tracked_user WHERE user_id=$1"
+        tracked_user_id = await self.db.fetchval(q, user_id)
         if tracked_user_id is None:
             return None
-        rows = await self.db.fetch(
-            "SELECT device_id, identity_key, signing_key, trust, deleted, "
-            "name FROM crypto_device WHERE user_id=$1",
-            user_id,
-        )
+        q = """
+        SELECT device_id, identity_key, signing_key, trust, deleted, name
+        FROM crypto_device WHERE user_id=$1
+        """
+        rows = await self.db.fetch(q, user_id)
         result = {}
         for row in rows:
             result[row["device_id"]] = DeviceIdentity(
@@ -438,12 +433,11 @@ class PgCryptoStore(CryptoStore, SyncStore):
         return result
 
     async def get_device(self, user_id: UserID, device_id: DeviceID) -> DeviceIdentity | None:
-        row = await self.db.fetchrow(
-            "SELECT identity_key, signing_key, trust, deleted, name "
-            "FROM crypto_device WHERE user_id=$1 AND device_id=$2",
-            user_id,
-            device_id,
-        )
+        q = """
+        SELECT identity_key, signing_key, trust, deleted, name FROM crypto_device
+        WHERE user_id=$1 AND device_id=$2
+        """
+        row = await self.db.fetchrow(q, user_id, device_id)
         if row is None:
             return None
         return DeviceIdentity(
@@ -459,9 +453,12 @@ class PgCryptoStore(CryptoStore, SyncStore):
     async def find_device_by_key(
         self, user_id: UserID, identity_key: IdentityKey
     ) -> DeviceIdentity | None:
+        q = """
+        SELECT device_id, signing_key, trust, deleted, name FROM crypto_device
+        WHERE user_id=$1 AND identity_key=$2
+        """
         row = await self.db.fetchrow(
-            "SELECT device_id, signing_key, trust, deleted, name "
-            "FROM crypto_device WHERE user_id=$1 AND identity_key=$2",
+            q,
             user_id,
             identity_key,
         )
@@ -500,32 +497,29 @@ class PgCryptoStore(CryptoStore, SyncStore):
             "name",
         ]
         async with self.db.acquire() as conn, conn.transaction():
-            await conn.execute(
-                "INSERT INTO crypto_tracked_user (user_id) VALUES ($1) "
-                "ON CONFLICT (user_id) DO NOTHING",
-                user_id,
-            )
+            q = """
+            INSERT INTO crypto_tracked_user (user_id) VALUES ($1) ON CONFLICT (user_id) DO NOTHING
+            """
+            await conn.execute(q, user_id)
             await conn.execute("DELETE FROM crypto_device WHERE user_id=$1", user_id)
             if self.db.scheme == Scheme.POSTGRES:
                 await conn.copy_records_to_table("crypto_device", records=data, columns=columns)
             else:
-                await conn.executemany(
-                    "INSERT INTO crypto_device (user_id, device_id, "
-                    "identity_key, signing_key, trust, deleted, name) "
-                    "VALUES ($1, $2, $3, $4, $5, $6, $7)",
-                    data,
-                )
+                q = """
+                INSERT INTO crypto_device (
+                    user_id, device_id, identity_key, signing_key, trust, deleted, name
+                ) VALUES ($1, $2, $3, $4, $5, $6, $7)
+                """
+                await conn.executemany(q, data)
 
     async def filter_tracked_users(self, users: list[UserID]) -> list[UserID]:
         if self.db.scheme in (Scheme.POSTGRES, Scheme.COCKROACH):
-            rows = await self.db.fetch(
-                "SELECT user_id FROM crypto_tracked_user WHERE user_id = ANY($1)", users
-            )
+            q = "SELECT user_id FROM crypto_tracked_user WHERE user_id = ANY($1)"
+            rows = await self.db.fetch(q, users)
         else:
             params = ",".join(["?"] * len(users))
-            rows = await self.db.fetch(
-                f"SELECT user_id FROM crypto_tracked_user WHERE user_id IN ({params})", *users
-            )
+            q = f"SELECT user_id FROM crypto_tracked_user WHERE user_id IN ({params})"
+            rows = await self.db.fetch(q, *users)
         return [row["user_id"] for row in rows]
 
     async def put_cross_signing_key(
